requirements.txt
absl-py==0.9.0
astor==0.8.1
atari-py==0.2.6
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
box2d-py==2.3.8
certifi==2019.11.28
cffi==1.11.5
chardet==3.0.4
cloudpickle==1.2.1
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
future==0.18.2
gast==0.3.3
google-pasta==0.1.8
graphsurgeon==0.4.1
grpcio==1.27.1
gym==0.15.4
h5py==2.10.0
html-testRunner==1.2.1
idna==2.8
imageio==2.6.1
importlib-metadata==1.3.0
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.2
Jinja2==2.10.3
joblib==0.14.1
jsonpatch==1.24
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
jupyterthemes==0.20.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lesscpy==0.14.0
Markdown==3.2
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
more-itertools==8.0.2
mpi4py==3.0.3
nbconvert==5.6.1
nbformat==5.0.3
notebook==6.0.2
numpy==1.18.1
opencv-python==4.1.2.30
opt-einsum==3.1.0
packaging==20.1
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
pkg-resources==0.0.0
pluggy==0.13.1
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.3
psutil==5.6.7
ptyprocess==0.6.0
py==1.8.1
pybullet==2.6.4
pycparser==2.19
pyglet==1.3.2
Pygments==2.5.2
pyinter==0.2.0
pyparsing==2.4.6
pyquaternion==0.9.5
PyRep==1.1
pyrsistent==0.15.7
pytest==5.3.5
python-dateutil==2.8.1
pytz==2019.3
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
rlbench==1.0
scipy==1.4.1
seaborn==0.8.1
Send2Trash==1.5.0
six==1.14.0
-e git+https://github.com/openai/spinningup.git@038665d62d569055401d91856abb287263096178#egg=spinup
stable-baselines==2.9.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tensorflow-probability==0.7.0
tensorrt==7.0.0.11
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
torch==1.3.1
torchfile==0.1.0
tornado==6.0.3
tqdm==4.42.0
traitlets==4.3.3
uff==0.6.5
urllib3==1.25.8
visdom==0.1.8.9
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
